requirements.txt
levenshtein
loguru
xmltodict
semver
packageurl-python
tqdm
requests
pyyaml